Combine SimpleCameraHandler and PinchCameraHandler
Rather than having separate handlers, just have a single one with a prop that decides which modality to use for moving (pinching or two-finger panning).
