+++ LookatExample.js
@@ -11,8 +11,7 @@ import {
   createIdentityMatrix
 } from './matrixMath';
 
-const SimpleCameraHandler = require('./SimpleCameraHandler');
-const PinchCameraHandler = require('./PinchCameraHandler');
+const LookatMatrixGestureHandler = require('./LookatMatrixGestureHandler');
 
 class LookatExample extends Component {
   state = {
@@ -22,7 +21,7 @@ class LookatExample extends Component {
   _onLookatChanged = matrix => {
     this.setState({
       lookat: matrix,
-      handler: 'pan',
+      moveGesture: 'pan',
     });
 
     // TODO: send lookat over data channel
@@ -33,43 +32,30 @@ class LookatExample extends Component {
   }
 
   render() {
-    let mainView = (
-      <View style={styles.container}>
-        <Text style={styles.text}>
-          <Text>|{this._format(this.state.lookat[0])}, {this._format(this.state.lookat[1])}, {this._format(this.state.lookat[2])}, {this._format(this.state.lookat[3])}|{'\n'}</Text>
-          <Text>|{this._format(this.state.lookat[4])}, {this._format(this.state.lookat[5])}, {this._format(this.state.lookat[6])}, {this._format(this.state.lookat[7])}|{'\n'}</Text>
-          <Text>|{this._format(this.state.lookat[8])}, {this._format(this.state.lookat[9])}, {this._format(this.state.lookat[10])}, {this._format(this.state.lookat[11])}|{'\n'}</Text>
-          <Text>|{this._format(this.state.lookat[12])}, {this._format(this.state.lookat[13])}, {this._format(this.state.lookat[14])}, {this._format(this.state.lookat[15])}|{'\n'}</Text>
-        </Text>
-        <Button onPress={() => _cameraHandler.reset()} title={'Reset!'} />
-        <Picker
-          style={styles.picker} 
-          selectedValue={this.state.handler}
-          onValueChange={value => this.setState({handler: value})}>
-          <Picker.Item label="Pan to Move" value="pan" />
-          <Picker.Item label="Pinch to Move" value="pinch" />
-        </Picker>
-      </View>
+    let _cameraHandler;
+    return (
+      <LookatMatrixGestureHandler
+        ref={(cameraHandler) => { _cameraHandler = cameraHandler; }}
+        moveGesture={this.state.moveGesture}
+        onLookatChanged={this._onLookatChanged}>
+        <View style={styles.container}>
+          <Text style={styles.text}>
+            <Text>|{this._format(this.state.lookat[0])}, {this._format(this.state.lookat[1])}, {this._format(this.state.lookat[2])}, {this._format(this.state.lookat[3])}|{'\n'}</Text>
+            <Text>|{this._format(this.state.lookat[4])}, {this._format(this.state.lookat[5])}, {this._format(this.state.lookat[6])}, {this._format(this.state.lookat[7])}|{'\n'}</Text>
+            <Text>|{this._format(this.state.lookat[8])}, {this._format(this.state.lookat[9])}, {this._format(this.state.lookat[10])}, {this._format(this.state.lookat[11])}|{'\n'}</Text>
+            <Text>|{this._format(this.state.lookat[12])}, {this._format(this.state.lookat[13])}, {this._format(this.state.lookat[14])}, {this._format(this.state.lookat[15])}|{'\n'}</Text>
+          </Text>
+          <Button onPress={() => _cameraHandler.reset()} title={'Reset'} />
+          <Picker
+            style={styles.picker} 
+            selectedValue={this.state.moveGesture}
+            onValueChange={value => this.setState({moveGesture: value})}>
+            <Picker.Item label="Pan to Move" value="pan" />
+            <Picker.Item label="Pinch to Move" value="pinch" />
+          </Picker>
+        </View>
+      </LookatMatrixGestureHandler>
     );
-
-    let _cameraHandler;    
-    if (this.state.handler === 'pinch') {
-      return (
-        <PinchCameraHandler
-          ref={(cameraHandler) => { _cameraHandler = cameraHandler; }}
-          onLookatChanged={this._onLookatChanged}>
-          {mainView}
-        </PinchCameraHandler>
-      );
-    } else {
-      return (
-        <SimpleCameraHandler
-          ref={(cameraHandler) => { _cameraHandler = cameraHandler; }}
-          onLookatChanged={this._onLookatChanged}>
-          {mainView}
-        </SimpleCameraHandler>
-      );      
-    }
   }
 }
