+++ LookatMatrixGestureHandler.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 
 import {
   PanGestureHandler,
+  PinchGestureHandler,
   State,
 } from 'react-native-gesture-handler';
 
@@ -24,14 +25,19 @@ class SimpleCameraHandler extends Component {
     pitch: 0.0,
     downPitch: 0.0,
     location: [0.0, 0.0, 0.0],
-    downLocation: [0.0, 0.0, 0.0]
+    downLocation: [0.0, 0.0, 0.0],
   }
 
   static propTypes = {
+    moveGesture: PropTypes.oneOf(['pan','pinch']),
     onLookatChanged: PropTypes.func,
   }
 
-  _onPanCamera = event => {
+  static defaultProps = {
+    moveGesture: 'pan',
+  }
+
+  _onPan = event => {
     let dx = event.nativeEvent.translationX;
     let dy = event.nativeEvent.translationY;
 
@@ -58,8 +64,32 @@ class SimpleCameraHandler extends Component {
     this.props.onLookatChanged && this.props.onLookatChanged(lookat);
   }
 
-  _onPanXYZ = event => {
+  _onPanMove = event => {
     let dy = -event.nativeEvent.translationY;
+
+    // update the navigation origin
+    let lookat = cloneMatrix(this.state.lookat);   
+    let location = [0.0, 0.0, 0.0]; 
+    location[0] = this.state.downLocation[0] + dy * lookat[0];
+    location[1] = this.state.downLocation[1] + dy * lookat[1];
+    location[2] = this.state.downLocation[2] + dy * lookat[2];
+
+    lookat[12] = location[0];
+    lookat[13] = location[1];
+    lookat[14] = location[2];
+
+    this.setState({
+      lookat: lookat,
+      location: location,
+    });
+
+    this.props.onLookatChanged && this.props.onLookatChanged(lookat);
+  }
+
+  _onPinchMove = event => {
+    let scale = event.nativeEvent.scale || 0.001;
+    let dy = scale > 1 ? (scale - 1.0) : -((1.0 / scale) - 1.0);
+    dy *= 100;
 
     // update the navigation origin
     let lookat = cloneMatrix(this.state.lookat);   
@@ -108,16 +138,20 @@ class SimpleCameraHandler extends Component {
   }
 
   render() {
-    return (
-      <PanGestureHandler
-        onGestureEvent={this._onPanCamera}
-        onHandlerStateChange={this._onStateChange}
-        minDist={10}
-        minPointers={1}
-        maxPointers={1}
-        avgTouches>
+    let moveGestureHandler;
+    if (this.props.moveGesture === 'pinch') {
+      moveGestureHandler = (
+        <PinchGestureHandler
+          onGestureEvent={this._onPinchMove}
+          onHandlerStateChange={this._onStateChange}
+          avgTouches>
+          {this.props.children}
+        </PinchGestureHandler>
+      );
+    } else {
+      moveGestureHandler = (
         <PanGestureHandler
-          onGestureEvent={this._onPanXYZ}
+          onGestureEvent={this._onPanMove}
           onHandlerStateChange={this._onStateChange}
           minDist={10}
           minPointers={2}
@@ -125,6 +159,18 @@ class SimpleCameraHandler extends Component {
           avgTouches>
           {this.props.children}
         </PanGestureHandler>
+      );
+    }
+
+    return (
+      <PanGestureHandler
+        onGestureEvent={this._onPan}
+        onHandlerStateChange={this._onStateChange}
+        minDist={10}
+        minPointers={1}
+        maxPointers={1}
+        avgTouches>
+        {moveGestureHandler}
       </PanGestureHandler>
     );
   }
